resnet50v1_5/requirements.txt
absl-py==0.8.1
apache-beam==2.19.0
astor==0.8.1
astunparse==1.6.3
avro-python3==1.9.2.1
blessings==1.7
bpython==0.18
cachetools==3.1.1
certifi==2019.11.28
chardet==3.0.4
crcmod==1.7
curtsies==0.3.1
dill==0.3.1.1
docopt==0.6.2
fastavro==0.21.24
fasteners==0.15
future==0.18.2
gast==0.3.3
google-api-core==1.16.0
google-api-python-client==1.8.0
google-apitools==0.5.28
google-auth==1.12.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
google-cloud-bigquery==1.17.1
google-cloud-bigtable==1.0.0
google-cloud-core==1.3.0
google-cloud-datastore==1.7.4
google-cloud-pubsub==1.0.2
google-pasta==0.1.8
google-resumable-media==0.4.1
googleapis-common-protos==1.51.0
greenlet==0.4.15
grpc-google-iam-v1==0.12.3
grpcio==1.26.0
h5py==2.10.0
hdfs==2.5.8
httplib2==0.12.0
idna==2.9
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
mock==2.0.0
monotonic==1.5
numpy==1.18.1
oauth2client==3.0.0
oauthlib==3.1.0
opt-einsum==3.2.0
pbr==5.4.4
protobuf==3.11.2
psutil==5.6.7
pyarrow==0.15.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydot==1.4.1
Pygments==2.6.1
pymongo==3.10.1
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
scipy==1.4.1
six==1.13.0
tensorboard==2.1.1
tensorflow==2.1.0
tensorflow-estimator==2.1.0
tensorflow-metadata==0.21.1
tensorflow-serving-api==2.1.0
tensorflow-transform==0.22.0.dev0
termcolor==1.1.0
tfx-bsl==0.21.4
typing==3.7.4.1
typing-extensions==3.7.4.1
uritemplate==3.0.1
urllib3==1.25.8
virtualenv==16.0.0
wcwidth==0.1.9
Werkzeug==0.16.0
wrapt==1.11.2
yapf==0.26.0
